requirements.txt

# Core web framework
streamlit>=1.28.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Web3 and blockchain
web3>=6.0.0
eth-account>=0.9.0
eth-utils>=2.0.0

# HTTP requests and APIs
requests>=2.31.0
aiohttp>=3.8.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Environment management
python-dotenv>=1.0.0

# Cryptography and security
cryptography>=41.0.0

# Utilities
pydantic>=2.0.0
typing-extensions>=4.7.0
asyncio
